Refactored some components
- Clean up file structure
- Updated manifest.json
- Update file name
