+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "A1b3rt",
+  "name": "A1b3rt: A better NYU Course Search",
   "icons": [
     {
       "src": "favicon.ico",
