+++ src/pages/FullCoursePage.tsx
